+++ lib/yard-sorbet/handlers/mixes_in_class_methods_handler.rb
@@ -12,16 +12,17 @@ module YARDSorbet
       handles method_call(:mixes_in_class_methods)
       namespace_only
 
-      @@mix_in_class_methods = T.let({}, T::Hash[String, String]) # rubocop:disable Style/ClassVars
+      @@mix_in_class_methods = T.let({}, T::Hash[String, T::Array[String]]) # rubocop:disable Style/ClassVars
 
-      sig { params(code_obj: String).returns(T.nilable(String)) }
+      sig { params(code_obj: String).returns(T.nilable(T::Array[String])) }
       def self.mixed_in_class_methods(code_obj)
         @@mix_in_class_methods[code_obj]
       end
 
       sig { void }
       def process
-        @@mix_in_class_methods[namespace.to_s] = statement.parameters(false)[0].source
+        @@mix_in_class_methods[namespace.to_s] ||= []
+        @@mix_in_class_methods.fetch(namespace.to_s) << statement.parameters(false)[0].source
       end
     end
   end