+++ lib/yard-sorbet/node_utils.rb
@@ -7,9 +7,9 @@ module YARDSorbet
     extend T::Sig
 
     # Command node types that can have type signatures
-    ATTRIBUTE_METHODS = T.let(Set[:attr, :attr_accessor, :attr_reader, :attr_writer].freeze, T::Set[Symbol])
+    ATTRIBUTE_METHODS = T.let(%i[attr attr_accessor attr_reader attr_writer].freeze, T::Array[Symbol])
     # Skip these method contents during BFS node traversal, they can have their own nested types via `T.Proc`
-    SKIP_METHOD_CONTENTS = T.let(Set[:params, :returns].freeze, T::Set[Symbol])
+    SKIP_METHOD_CONTENTS = T.let(%i[params returns].freeze, T::Array[Symbol])
     # Node types that can have type signatures
     SigableNode = T.type_alias { T.any(YARD::Parser::Ruby::MethodDefinitionNode, YARD::Parser::Ruby::MethodCallNode) }
     private_constant :ATTRIBUTE_METHODS, :SKIP_METHOD_CONTENTS, :SigableNode